+++ apx_core/src/http_signatures/verify.rs
@@ -175,14 +175,11 @@ pub fn verify_http_signature(
         return Err(VerificationError::DigestMismatch);
     };
     let signature = base64::decode(&signature_data.signature)?;
-    let is_valid_signature = verify_rsa_sha256_signature(
+    verify_rsa_sha256_signature(
         signer_key,
         signature_data.message.as_bytes(),
         &signature,
-    );
-    if !is_valid_signature {
-        return Err(VerificationError::InvalidSignature);
-    };
+    ).map_err(|_| VerificationError::InvalidSignature)?;
     Ok(())
 }